Remade Shape game to Finish
I got carried away with the observer pattern and needed to trash every thing from the original version. Also, learned about the physics matrix in unity which solves friendly unit interaction problem from the original game.
Added a base spawner with 3 locations that can be cycled through from which any of the 3 units can be spawned from.
Also, this branch needs rebased as the main, I have no plan to implement the new input system in this project.