Import Mirror
Haven't gotten all functionality working with mirror's framework, but connecting 2 instances locally works with unique player controls on the spawners, but the instantiated prefabs do not sync over the network. Need to read up on mirror's solution to this. Initial look appears to be their spawning system. Pushing to work remotely.