Needs tested online
Added functionality to move location of spawn and choose unit to spawn, working perfectly on host and client locally. Would like to make the fire cmd call a part of the unit selection so you choose a location then when you hit the button to select a unit it is spawn on press so you do not have to click afterwards. To make the game more fast-paced, a more circular map may work better to allow for more possible collisions and some more distance could make the game more strategic and less reaction time based, though I don't think the reaction time dependence should be completely removed.