Refactored Code
Extracted a few functions, made spawn unit take an integer to determine the unit to spawn. Removed the weapon class and replaced it with the spawner transforms. moved camera to be top down, though from the side a vertical layout for the units may be the way to go (tower of mages).